Task 1: Style an HTML page using CSS..
Open the styles.css file.

Add a CSS rule for the body element that sets the background color to #E0E0E2.

Add a CSS rule for the h1 element that sets the text color to: #721817.

Add a CSS rule for the h2 element that sets the text color to: #721817.

Add a CSS rule for the center-text CSS class that aligns the text to center.

Add a CSS rule for the HTML element with the id logo. Set its left and right margins to auto and changes its display to a block element.

Add a CSS rule for all span elements that are children of h2 elements that sets the text color to #FA9F42 and its font size to 0.75em.

Add a CSS rule for the HTML element with the id copyright. Set its top padding to 12 pixels and its font size to 0.75em.